fix bootstrap
