+++ libs/adapters/src/daemon.ts
@@ -1,6 +1,48 @@
-import { logger } from '@hicommonwealth/core';
+import { RedisNamespaces, logger } from '@hicommonwealth/core';
+import { CacheDecorator, KeyFunction } from './redis';
 
 const log = logger().getLogger(__filename);
+
+export class Activity<T extends (...args: any[]) => any> {
+  queryWithCache: T;
+  queryWithCacheOverride: T;
+
+  constructor(
+    private label: string,
+    private query: T,
+    private cacheKey: KeyFunction<T>,
+    private cacheDuration: number,
+    private cacheNamespace: RedisNamespaces,
+    private cacheDecorator: CacheDecorator,
+  ) {
+    this.queryWithCache = this.cacheWrapHelper(false);
+    this.queryWithCacheOverride = this.cacheWrapHelper(true);
+  }
+
+  cacheWrapHelper(override: boolean) {
+    return this.cacheDecorator.cacheWrap(
+      override,
+      this.query,
+      this.cacheKey,
+      this.cacheDuration,
+      this.cacheNamespace,
+    ) as unknown as T;
+  }
+
+  async startTask(...args: any) {
+    try {
+      const jobId = daemon.startTask(
+        this.label,
+        async () => await this.queryWithCacheOverride(...args),
+        this.cacheDuration,
+      );
+      return jobId;
+    } catch (err) {
+      console.error(err);
+      return;
+    }
+  }
+}
 
 type DaemonTask = () => void;
 export class Daemons {
