+++ libs/adapters/src/hot-shots/index.ts
@@ -3,7 +3,7 @@ import { StatsD } from 'hot-shots';
 
 export const HotShotsStats = (): Stats => {
   const log = logger().getLogger(__filename);
-  const client = new StatsD({
+  let client = new StatsD({
     globalTags: { env: process.env.NODE_ENV || 'development' },
     errorHandler: (error) => {
       log.error('Caught statsd socket error', error);
@@ -11,18 +11,27 @@ export const HotShotsStats = (): Stats => {
   });
   return {
     name: 'HotShotStats',
-    dispose: () => {
-      client.close((error) => log.error(error.message, error));
-      return Promise.resolve();
-    },
-    histogram: (key, value, tags) => client.histogram(key, value, tags),
-    set: (key, value) => client.set(key, value),
-    increment: (key, tags) => client.increment(key, tags),
-    incrementBy: (key, value, tags) => client.increment(key, value, tags),
-    decrement: (key, tags) => client.decrement(key, tags),
-    decrementBy: (key, value, tags) => client.decrement(key, value, tags),
-    on: (key) => client.gauge(key, 1),
-    off: (key) => client.gauge(key, 0),
-    timing: (key, duration, tags) => client.timing(key, duration, tags),
+    dispose: () =>
+      new Promise((resolve) => {
+        client &&
+          client.close((error) => {
+            error && log.error(error.message, error);
+            resolve();
+          });
+        client = undefined;
+      }),
+    histogram: (key, value, tags) =>
+      client && client.histogram(key, value, tags),
+    set: (key, value) => client && client.set(key, value),
+    increment: (key, tags) => client && client.increment(key, tags),
+    incrementBy: (key, value, tags) =>
+      client && client.increment(key, value, tags),
+    decrement: (key, tags) => client && client.decrement(key, tags),
+    decrementBy: (key, value, tags) =>
+      client && client.decrement(key, value, tags),
+    on: (key) => client && client.gauge(key, 1),
+    off: (key) => client && client.gauge(key, 0),
+    timing: (key, duration, tags) =>
+      client && client.timing(key, duration, tags),
   };
 };
