+++ libs/adapters/src/redis/redisCache.ts
@@ -1,4 +1,4 @@
-import { logger, type RedisNamespaces } from '@hicommonwealth/core';
+import { ILogger, logger, type RedisNamespaces } from '@hicommonwealth/core';
 import {
   ConnectionTimeoutError,
   createClient,
@@ -6,8 +6,7 @@ import {
   SocketClosedUnexpectedlyError,
 } from 'redis';
 import type Rollbar from 'rollbar';
-
-const log = logger().getLogger(__filename);
+import { timeoutPromise } from '../utils';
 
 export function redisRetryStrategy(retries: number) {
   if (retries > 5) {
@@ -29,8 +28,10 @@ export class RedisCache {
   private _initialized = false;
   private _client;
   private _rollbar?: Rollbar;
+  private _log?: ILogger;
 
   constructor(rollbar?: Rollbar) {
+    this._log = logger().getLogger(__filename);
     this._rollbar = rollbar;
   }
 
@@ -44,13 +45,13 @@ export class RedisCache {
    */
   public async init(redis_url: string) {
     if (!redis_url) {
-      log.warn(
+      this._log.warn(
         'Redis Url is undefined. Some services (e.g. chat) may not be available.',
       );
       this._initialized = false;
       return;
     }
-    log.info(`Connecting to Redis at: ${redis_url}`);
+    this._log.info(`Connecting to Redis at: ${redis_url}`);
 
     if (!this._client) {
       const redisOptions = {};
@@ -74,37 +75,37 @@ export class RedisCache {
     this._client.on('error', (err) => {
       if (err instanceof ConnectionTimeoutError) {
         const msg = `RedisCache connection to ${redis_url} timed out!`;
-        log.error(msg);
+        this._log.error(msg);
         this._rollbar?.error(
           'RedisCache max connection retries exceeded! RedisCache client shutting down!',
         );
       } else if (err instanceof ReconnectStrategyError) {
         const msg =
           'RedisCache max connection retries exceeded! RedisCache client shutting down!';
-        log.fatal(msg);
+        this._log.fatal(msg);
         this._rollbar?.critical(msg);
       } else if (err instanceof SocketClosedUnexpectedlyError) {
         const msg = 'RedisCache socket closed unexpectedly';
-        log.error(msg);
+        this._log.error(msg);
         this._rollbar?.error(msg);
       } else {
         const msg = 'RedisCache unknown connection error:';
-        log.error(msg, err);
+        this._log.error(msg, err);
         this._rollbar?.critical(msg, err);
       }
     });
 
     this._client.on('ready', () => {
       this._initialized = !!this._client.isOpen;
-      log.info(`RedisCache connection ready ${this._initialized}`);
+      this._log.info(`RedisCache connection ready ${this._initialized}`);
     });
     this._client.on('reconnecting', () => {
       this._initialized = !!this._client.isOpen;
-      log.info(`RedisCache reconnecting ${this._initialized}`);
+      this._log.info(`RedisCache reconnecting ${this._initialized}`);
     });
     this._client.on('end', () => {
       this._initialized = !!this._client.isOpen;
-      log.info(`RedisCache disconnected ${this._initialized}`);
+      this._log.info(`RedisCache disconnected ${this._initialized}`);
     });
 
     if (!this._client.isOpen) {
@@ -116,7 +117,9 @@ export class RedisCache {
 
   private initialized() {
     if (!this._initialized) {
-      log.warn('Redis client is not initialized. Run RedisCache.init() first!');
+      this._log.warn(
+        'Redis client is not initialized. Run RedisCache.init() first!',
+      );
       return false;
     }
     return true;
@@ -163,7 +166,7 @@ export class RedisCache {
       }
     } catch (e) {
       const msg = `An error occurred while setting the following key value pair '${namespace} ${key}: ${value}'`;
-      log.error(msg, e);
+      this._log.error(msg, e);
       this._rollbar?.error(msg, e);
       return false;
     }
@@ -183,7 +186,7 @@ export class RedisCache {
       }
     } catch (e) {
       const msg = `An error occurred while getting the following key '${key}'`;
-      log.error(msg, e);
+      this._log.error(msg, e);
       this._rollbar?.error(msg, e);
     }
   }
@@ -229,7 +232,7 @@ export class RedisCache {
           const msg =
             `Error occurred while setting multiple keys ` +
             `${transaction ? 'in a transaction' : 'in a pipeline'}`;
-          log.error(msg, e);
+          this._log.error(msg, e);
           this._rollbar?.error(msg, e);
           return false;
         }
@@ -238,7 +241,7 @@ export class RedisCache {
           return await this._client.MSET(transformedData);
         } catch (e) {
           const msg = 'Error occurred while setting multiple keys';
-          log.error(msg, e);
+          this._log.error(msg, e);
           this._rollbar?.error(msg, e);
           return false;
         }
@@ -265,7 +268,7 @@ export class RedisCache {
         }, {});
       } catch (e) {
         const msg = 'An error occurred while getting many keys';
-        log.error(msg, e);
+        this._log.error(msg, e);
         this._rollbar?.error(msg, e);
         return false;
       }
@@ -299,7 +302,7 @@ export class RedisCache {
       }
     } catch (e) {
       const msg = 'An error occurred while fetching the namespace keys';
-      log.error(msg, e);
+      this._log.error(msg, e);
       this._rollbar?.error(msg, e);
       return false;
     }
@@ -340,17 +343,17 @@ export class RedisCache {
           try {
             const resp = await this._client.del(key);
             count += resp;
-            log.trace(`deleted key ${key} ${resp} ${count}`);
+            this._log.trace(`deleted key ${key} ${resp} ${count}`);
           } catch (err) {
-            log.trace(`error deleting key ${key}`);
-            log.trace(err);
+            this._log.trace(`error deleting key ${key}`);
+            this._log.trace(err);
           }
         }
       }
       return count;
     } catch (e) {
       const msg = `An error occurred while deleting a all keys in the ${namespace} namespace`;
-      log.error(msg, e);
+      this._log.error(msg, e);
       this._rollbar?.error(msg, e);
       return false;
     }
@@ -367,7 +370,7 @@ export class RedisCache {
       }
     } catch (e) {
       const msg = `An error occurred while deleting the following key: ${finalKey}`;
-      log.error(msg, e);
+      this._log.error(msg, e);
       this._rollbar?.error(msg, e);
       return 0;
     }
@@ -376,4 +379,11 @@ export class RedisCache {
   public get client(): typeof this._client {
     return this._client;
   }
+}
+
+export async function connectToRedis(redisCache: RedisCache) {
+  const REDIS_URL = process.env.REDIS_URL || 'redis://localhost:6379';
+  if (!REDIS_URL) throw new Error('REDIS_URL not set');
+  await Promise.race([timeoutPromise(10000), redisCache.init(REDIS_URL)]);
+  if (!redisCache.isInitialized) throw new Error('Redis Cache not initialized');
 }
