+++ libs/core/src/command.ts
