+++ libs/core/src/errors.ts
