+++ libs/core/src/index.ts
@@ -1,3 +1,5 @@
+export * from './command';
+export * from './errors';
 export * from './ports';
 export * from './requirements-types';
 export * from './shared-types';
