+++ libs/model/src/community/CreateCommunity.command.ts
@@ -1,3 +1,4 @@
+import type { CommandMetadata } from '@hicommonwealth/core';
 import {
   ChainBase,
   ChainType,
@@ -6,7 +7,6 @@ import {
 import { z } from 'zod';
 import { models } from '../database';
 import type { CommunityAttributes } from '../models';
-import type { CommandMetadata } from '../types';
 import { checkIconSize } from '../utils/checkIconSize';
 import { ALL_COMMUNITIES } from '../utils/constants';
