+++ libs/model/src/index.ts
@@ -1,5 +1,4 @@
 // Core Domain
-export * from './command';
 export * as Comment from './comment';
 export * as Community from './community';
 export * as Reaction from './reaction';
@@ -8,7 +7,5 @@ export * as User from './user';
 
 // Internals - Should not be exported once we finish the migrations to models
 export * from './database';
-export * from './errors';
 export * from './models';
-export * from './types';
 export * from './utils/abi';
