+++ packages/commonwealth/server-test.ts
@@ -1,10 +1,7 @@
 /* eslint-disable dot-notation */
 import {
-  CustomRequest,
+  CacheDecorator,
   RedisCache,
-  ServerError,
-  cacheDecorator,
-  lookupKeyDurationInReq,
   setupErrorHandlers,
 } from '@hicommonwealth/adapters';
 import { logger } from '@hicommonwealth/core';
@@ -12,7 +9,6 @@ import { models } from '@hicommonwealth/model';
 import bodyParser from 'body-parser';
 import SessionSequelizeStore from 'connect-session-sequelize';
 import cookieParser from 'cookie-parser';
-import type { Express } from 'express';
 import express from 'express';
 import session from 'express-session';
 import http from 'http';
@@ -115,77 +111,6 @@ const setupServer = () => {
   server.on('listening', onListen);
 };
 
-export enum CACHE_ENDPOINTS {
-  BROKEN_5XX = '/cachedummy/broken5xx',
-  BROKEN_4XX = '/cachedummy/broken4xx',
-  JSON = '/cachedummy/json',
-  TEXT = '/cachedummy/text',
-  CUSTOM_KEY_DURATION = '/cachedummy/customKeyDuration',
-}
-
-export const setupCacheTestEndpoints = (appAttach: Express) => {
-  // /cachedummy endpoint for testing
-  appAttach.get(
-    CACHE_ENDPOINTS.BROKEN_4XX,
-    cacheDecorator.cacheMiddleware(3),
-    async (req, res) => {
-      log.info(`${CACHE_ENDPOINTS.BROKEN_4XX} called`);
-      res.status(400).json({ message: 'cachedummy 400 response' });
-    },
-  );
-
-  appAttach.get(
-    CACHE_ENDPOINTS.JSON,
-    cacheDecorator.cacheMiddleware(3),
-    async (req, res) => {
-      log.info(`${CACHE_ENDPOINTS.JSON} called`);
-      res.json({ message: 'cachedummy response' });
-    },
-  );
-
-  appAttach.post(
-    CACHE_ENDPOINTS.CUSTOM_KEY_DURATION,
-    (req: CustomRequest, res, next) => {
-      log.info(`${CACHE_ENDPOINTS.CUSTOM_KEY_DURATION} called`);
-      const body = req.body;
-      if (!body || !body.duration || !body.key) {
-        return next();
-      }
-      req.cacheKey = body.key;
-      req.cacheDuration = body.duration;
-      return next();
-    },
-    cacheDecorator.cacheMiddleware(3, lookupKeyDurationInReq),
-    async (req, res) => {
-      res.json(req.body);
-    },
-  );
-
-  // Uncomment the following lines if you want to use the /cachedummy/json route
-  // app.post('/cachedummy/json', cacheDecorator.cacheInvalidMiddleware(3), async (req, res) => {
-  //   res.json({ 'message': 'cachedummy response' });
-  // });
-
-  appAttach.get(
-    CACHE_ENDPOINTS.TEXT,
-    cacheDecorator.cacheMiddleware(3),
-    async function cacheTextEndpoint(req, res) {
-      log.info(`${CACHE_ENDPOINTS.TEXT} called`);
-      res.send('cachedummy response');
-    },
-  );
-
-  appAttach.get(
-    CACHE_ENDPOINTS.BROKEN_5XX,
-    cacheDecorator.cacheMiddleware(3),
-    async (req, res, next) => {
-      log.info(`${CACHE_ENDPOINTS.BROKEN_5XX} called`);
-      const err = new Error('route error');
-      return next(new ServerError('broken route', err));
-    },
-  );
-};
-
 const banCache = new BanCache(models);
 const redisCache = new RedisCache();
 const globalActivityCache = new GlobalActivityCache(models, redisCache);
@@ -203,8 +128,7 @@ setupAPI(
   databaseValidationService,
   redisCache,
 );
-setupCosmosProxy(app, models);
-setupCacheTestEndpoints(app);
+setupCosmosProxy(app, models, new CacheDecorator());
 
 const rollbar = new Rollbar({
   accessToken: ROLLBAR_SERVER_TOKEN,
