+++ packages/commonwealth/server/workers/commonwealthConsumer/commonwealthConsumer.ts
@@ -1,19 +1,24 @@
 import {
+  HotShotsStats,
   RabbitMQController,
   RabbitMQSubscription,
   RascalConfigServices,
   RascalSubscriptions,
   ServiceConsumer,
   ServiceKey,
+  TypescriptLoggingLogger,
   getRabbitMQConfig,
   startHealthCheckLoop,
 } from '@hicommonwealth/adapters';
-import { logger } from '@hicommonwealth/core';
+import { logger, stats } from '@hicommonwealth/core';
 import { models } from '@hicommonwealth/model';
 import type { BrokerConfig } from 'rascal';
 import Rollbar from 'rollbar';
 import { RABBITMQ_URI, ROLLBAR_ENV, ROLLBAR_SERVER_TOKEN } from '../../config';
 import { processSnapshotMessage } from './messageProcessors/snapshotConsumer';
+
+const log = logger(TypescriptLoggingLogger()).getLogger(__filename);
+stats(HotShotsStats());
 
 let isServiceHealthy = false;
 
@@ -34,8 +39,6 @@ startHealthCheckLoop({
 // to the CommonwealthConsumer and you want to ensure that the CommonwealthConsumer is
 // properly handling/processing those messages. Using the script is rarely necessary in
 // local development.
-
-const log = logger().getLogger(__filename);
 
 export async function setupCommonwealthConsumer(): Promise<ServiceConsumer> {
   const rollbar = new Rollbar({
