+++ packages/commonwealth/server/workers/evmChainEvents/startEvmPolling.ts
@@ -1,8 +1,9 @@
+import { TypescriptLoggingLogger } from '@hicommonwealth/adapters';
 import { logger } from '@hicommonwealth/core';
 import { rollbar } from '../../util/rollbar';
 import { processChainNode, scheduleNodeProcessing } from './nodeProcessing';
 
-const log = logger().getLogger(__filename);
+const log = logger(TypescriptLoggingLogger()).getLogger(__filename);
 
 /**
  * Starts an infinite loop that periodically fetches and parses blocks from
